backend/requirements.txt
# django
Django==4.0.1

# drf
djangorestframework==3.13.1
Markdown==3.3.6
django-crispy-forms==1.14.0
django-filter==21.1

# envs
python-environ==0.4.54

# postgres
psycopg2==2.8.6

# cors
django-cors-headers==3.11.0
